backend/requirements.txt
aniso8601==9.0.1
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
filelock==3.6.0
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-SocketIO==4.3.1
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.1.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.3
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
socketIO-client==0.7.2
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
transformers==4.17.0
typing-extensions==4.0.1
urllib3==1.26.9
websocket-client==1.2.1
websockets==10.1
Werkzeug==2.0.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
